refactor: update import paths and improve code organization across multiple files
