+++ frontend/package.json
@@ -47,6 +47,8 @@
     "dotenv-expand": "^12.0.2",
     "eslint": "^9.23.0",
     "eslint-config-prettier": "^10.1.2",
+    "eslint-import-resolver-typescript": "^4.4.4",
+    "eslint-plugin-import": "^2.32.0",
     "eslint-plugin-prettier": "^5.3.1",
     "eslint-plugin-sonarjs": "^3.0.2",
     "jasmine-core": "~5.6.0",
