+++ frontend/src/adapters/http/agile-wheel-backend.http.ts
@@ -1,8 +1,9 @@
-import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
+import { Injectable } from '@angular/core';
+
 import { Observable } from 'rxjs';
 
-import { environment } from '@env/environment';
+import { environment } from 'environments/environment';
 
 @Injectable({
   providedIn: 'root',
