+++ frontend/src/application/ports/activity-flow-use-case-port.ts
@@ -1,4 +1,5 @@
 import { InjectionToken } from '@angular/core';
+
 import { Activity, Participant, DimensionWithScores } from 'domain/model';
 
 /**
