+++ frontend/src/application/ports/close-activity-use-case-port.ts
@@ -1,4 +1,5 @@
 import { InjectionToken } from '@angular/core';
+
 import { Activity, Participant } from 'domain/model';
 
 /**
