+++ frontend/src/application/ports/create-activity-use-case-port.ts
@@ -1,5 +1,7 @@
 import { InjectionToken } from '@angular/core';
+
 import { Observable } from 'rxjs';
+
 import {
   CreateActivityRequest,
   CreateActivityResponse,
