+++ frontend/src/application/use-cases/activity-stream.usecase.ts
@@ -1,8 +1,9 @@
 import { Injectable } from '@angular/core';
+
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
 
-import { AgileWheelBackEndWS } from 'adapters/agile-wheel-backend.ws';
+import { AgileWheelBackEndWS } from 'adapters/web-socket/agile-wheel-backend.ws';
 import { ActivityStreamMessage } from 'application/dtos/activity-stream.dto';
 import { ActivityStreamUseCasePort } from 'application/ports/activity-stream-use-case-port';
