+++ frontend/src/presentations/pages/result/chart-result/chart-result.component.ts
@@ -1,6 +1,6 @@
 import { Component, ElementRef, ViewChild, AfterViewInit } from '@angular/core';
 
-import * as echarts from 'echarts/core';
+import { BarChart } from 'echarts/charts';
 import {
   TitleComponent,
   TooltipComponent,
@@ -8,7 +8,7 @@ import {
   PolarComponent,
   GraphicComponent,
 } from 'echarts/components';
-import { BarChart } from 'echarts/charts';
+import * as echarts from 'echarts/core';
 import { CanvasRenderer } from 'echarts/renderers';
 
 echarts.use([
